postcards/solution/backend/requirements.txt

fastapi
pydantic
openai
requests
python-dotenv
uvicorn